src/WhatsappIntegration/requirements.txt
flask==2.3.3
twilio==8.5.0
python-dotenv==1.0.0
openai==1.2.4
requests==2.31.0
gunicorn==21.2.0 
supabase
python-jose
